requirements.txt
fastapi>=0.115.0
uvicorn>=0.25.0
openai>=1.0.0
pinecone-client>=3.0.0
sentence-transformers>=0.5.1
jinja2>=3.0
python-multipart>=0.0.7